+++ Octokit/Models/AuthorizationUpdate.cs
@@ -1,0 +1,23 @@
+using System.Diagnostics.CodeAnalysis;
+
+namespace Octokit
+{
+    public class AuthorizationUpdate
+    {
+        /// <summary>
+        /// Replace scopes with this list.
+        /// </summary>
+        [SuppressMessage("Microsoft.Performance", "CA1819:PropertiesShouldNotReturnArrays", Justification = "Special type of model object that only updates none-null fields.")]
+        public string[] Scopes { get; set; }
+
+        /// <summary>
+        /// Notes about this particular <see cref="Authorization"/>.
+        /// </summary>
+        public string Note { get; set; }
+
+        /// <summary>
+        /// A url for more information about notes.
+        /// </summary>
+        public string NoteUrl { get; set; }
+    }
+}
