+++ Octokit/Models/EmailAddress.cs
@@ -1,0 +1,11 @@
+namespace Octokit
+{
+    public class EmailAddress
+    {
+        public string Email { get; set; }
+
+        public bool Verified { get; set; }
+
+        public bool Primary { get; set; }
+    }
+}
