+++ Octokit/Models/ReadmeResponse.cs
@@ -1,0 +1,11 @@
+namespace Octokit
+{
+    internal class ReadmeResponse
+    {
+        public string Content { get; set; }
+        public string Name { get; set; }
+        public string HtmlUrl { get; set; }
+        public string Url { get; set; }
+        public string Encoding { get; set; }
+    }
+}
