+++ Octokit/Models/Release.cs
@@ -1,0 +1,23 @@
+using System;
+using System.Diagnostics.CodeAnalysis;
+
+namespace Octokit
+{
+    public class Release
+    {
+        public string Url { get; set; }
+        public string HtmlUrl { get; set; }
+        public string AssetsUrl { get; set; }
+        public string UploadUrl { get; set; }
+        public int Id { get; set; }
+        public string TagName { get; set; }
+        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "Commitish")]
+        public string TargetCommitish { get; set; }
+        public string Name { get; set; }
+        public string Body { get; set; }
+        public bool Draft { get; set; }
+        public bool Prerelease { get; set; }
+        public DateTimeOffset CreatedAt { get; set; }
+        public DateTimeOffset PublishedAt { get; set; }
+    }
+}
