+++ Octokit/Models/SshKeyUpdate.cs
@@ -1,0 +1,15 @@
+namespace Octokit
+{
+    public class SshKeyUpdate
+    {
+        /// <summary>
+        /// The SSH Key
+        /// </summary>
+        public string Key { get; set; }
+
+        /// <summary>
+        /// The title of the SSH key
+        /// </summary>
+        public string Title { get; set; }
+    }
+}
